// packages/state-management-frontend/src/app/business/models/business.interface.ts
export interface Business {
  name: string;
  email: string;
  password: string;
  classification: string;
  address: string;
  longitude: string;
  latitude: string;
  contact: string;
  picture: string | null;
  bankAccountNumber: string | null;
  bankName: string | null;
  bankAccountType: string | null;
  fullname: string;
  documentId: string;
}